refactor(plugins/kuma): refactor custom block plugins specific to Kuma
Remove separate `tip.rb` and `warning.rb` plugins, which were just copies of
the same code with only difference in a block name. Now the code is the same
and just registered for different tags.

It also correctly maps different class names between Kong and Kuma docs.
